Remove unused files and update dependencies
Unused files and workflows were deleted to have a cleaner and more manageable project structure. As for the updates, the package.json was updated for its dependencies to have the latest version. Also, removed the unnecessary frozen string literal from the Gemfile and fixed the directory mapping from the .idea/vcs.xml and .idea/deployment.xml. Minor changes were also made to several other files to improve the coding and implementation logic. These changes were made to keep the project up-to-date and ensure that it is using the most recent versions of its dependencies, thus leveraging the latest features and bug fixes. Also, deleting unnecessary files helps to maintain a more uncluttered and manageable codebase.