Update dependencies and improve data handling
Upgraded dependent libraries in 'package.json' for better performance and security. So the new version of '@nuxt/content', '@nuxt/devtools', '@types/node', and 'nuxt' are installed. The '@nuxthq/ui' is replaced with '@nuxt/ui'. In 'nuxt.config.ts', the '@nuxthq/ui' module also is replaced with '@nuxt/ui'. Also updated the data fetching and handling in 'app.vue', replacing the old API URL with a new one, 'https://getthis.page/api', that returns cleaner and easier-to-handle data. Furthermore, adjustments were made to integrate and properly display the fetched data on the page. The changes promote better performance, security, maintainability, and user experience.